lab5/lab5_py/requirements.txt
simple-hexdump